Disable Rubocop Metrics/ClassLength cop for TravelTime::Client.
It's true that in general long classes are an anti-pattern, but in this case there's value in having all the API methods into a single class and splitting them into different files seems premature at the moment.
Also, we're just slightly over the default limit (101/100), so it would be a bit forced as well.